requirements.txt
flask==3.0.0
openai==1.12.0
supabase==2.3.0
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.9
langchain-openai==0.0.8
faiss-cpu==1.7.4
gunicorn==21.2.0
